pyproject.toml
[project]
name = "openknow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb",
    "openai",
    "mcp[cli]",
    "pyyaml",
    "rich",
    "Pillow",
    "chromadb>=1.5.2",
    "sentence-transformers>=5.2.3",
]

[project.optional-dependencies]
local = [
    "sentence-transformers",
]

[project.scripts]
openknow = "main:cli_main"
